tools/respeaker_mic_tunning/requirements.txt
pyusb
PySimpleGUI

